requirements.txt
appdirs==1.4.4
attrs==19.3.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
hidpidaemon==18.4.6
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
jsonschema==3.2.0
kernelstub==3.1.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
pop-transition==1.1.2
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
pydbus==0.6.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt===2.1.2pop0-1587756471-20.04-cd2988e
python-debian===0.1.36ubuntu1
python-dotenv==0.17.0
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
repoman==1.2.2
requests==2.22.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sessioninstaller==0.0.0
simplejson==3.16.0
six==1.14.0
systemd-python==234
texttable==1.6.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8
virtualenv==20.4.4
wadllib==1.3.3
websocket-client==0.53.0
xkit==0.0.0
zipp==1.0.0
